+++ packages/sources/por-indexer/src/endpoint/index.ts
@@ -1,0 +1,1 @@
+export { endpoint as balance } from './balance'