+++ packages/sources/the-network-firm/CHANGELOG.md
@@ -1,1 +1,14 @@
 # @chainlink/the-network-firm-adapter
+
+## 1.1.0
+
+### Minor Changes
+
+- d127d1287: Added the 'backed' endpoint
+- 52b58f481: Added the USDR endpoint
+
+## 1.0.0
+
+### Major Changes
+
+- a00bf5081: Initial version of the adapter. Forked from the armanino adapter.